Added support for GitVersion@5
