+++ src/AzurePipelinesToGitHubActionsConverter.Tests/StepsTests.cs
@@ -211,8 +211,8 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
 ";
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }   
-        
+        }
+
         [TestMethod]
         public void CheckOutIndividualStepTest()
         {
@@ -566,8 +566,8 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
 ";
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        } 
-        
+        }
+
         [TestMethod]
         public void AzurePowershellWithOtherVersionIndividualStepTest()
         {
@@ -602,7 +602,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
-        
+
         [TestMethod]
         public void AzurePowershellWithOtherVersionAndTargetAzurePsIndividualStepTest()
         {
@@ -637,7 +637,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
-        
+
         [TestMethod]
         public void AzurePowershellWithOtherVersionAndCustomTargetAzurePsIndividualStepTest()
         {
@@ -816,8 +816,8 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
 ";
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-        }  
-        
+        }
+
         [TestMethod]
         public void DownloadWithPatternsIndividualStepTest()
         {
@@ -981,7 +981,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
-        
+
 
         [TestMethod]
         public void DotNetCoreCLIPushIndividualStepTest()
@@ -1173,7 +1173,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
-        
+
         [TestMethod]
         public void FunctionalTestIndividualStepTest()
         {
@@ -1261,6 +1261,32 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
         }
 
         [TestMethod]
+        public void GitVersionIndividualStepTest()
+        {
+            //Arrange
+            Conversion conversion = new Conversion();
+            string yaml = @"
+- task: GitVersion@5
+  name: gitVersion
+  displayName: 'Evaluate Next Version'
+  inputs:
+    runtime: 'core'
+    configFilePath: 'GitVersion.yml'
+";
+
+            //Act
+            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+
+            //Assert
+            string expected = @"
+- name: Evaluate Next Version
+  uses: gittools/actions/gitversion/execute@v0.9.7
+";
+            expected = UtilityTests.TrimNewLines(expected);
+            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        }
+
+        [TestMethod]
         public void HugoIndividualStepTest()
         {
             //Arrange
@@ -1290,42 +1316,42 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
 
-//        [TestMethod]
-//        public void IISWebManagementStepTest()
-//        {
-//            //Arrange
-//            Conversion conversion = new Conversion();
-//            string yaml = @"
-//              - task: IISWebAppManagementOnMachineGroup@0
-//                inputs:
-//                  IISDeploymentType: 'IISWebsite'
-//                  ActionIISWebsite: 'CreateOrUpdateWebsite'
-//                  WebsiteName: 'Spark'
-//                  WebsitePhysicalPath: '%SystemDrive%\inetpub\wwwroot'
-//                  WebsitePhysicalPathAuth: 'WebsiteUserPassThrough'
-//                  AddBinding: true
-//                  CreateOrUpdateAppPoolForWebsite: true
-//                  ConfigureAuthenticationForWebsite: true
-//                  AppPoolNameForWebsite: 'Spark'
-//                  DotNetVersionForWebsite: 'v4.0'
-//                  PipeLineModeForWebsite: 'Integrated'
-//                  AppPoolIdentityForWebsite: 'ApplicationPoolIdentity'
-//                  AnonymousAuthenticationForWebsite: true
-//                  WindowsAuthenticationForWebsite: false
-//                  protocol: 'http' 
-//                  iPAddress: 'All Unassigned'
-//                  port: '80'
-//";
+        //        [TestMethod]
+        //        public void IISWebManagementStepTest()
+        //        {
+        //            //Arrange
+        //            Conversion conversion = new Conversion();
+        //            string yaml = @"
+        //              - task: IISWebAppManagementOnMachineGroup@0
+        //                inputs:
+        //                  IISDeploymentType: 'IISWebsite'
+        //                  ActionIISWebsite: 'CreateOrUpdateWebsite'
+        //                  WebsiteName: 'Spark'
+        //                  WebsitePhysicalPath: '%SystemDrive%\inetpub\wwwroot'
+        //                  WebsitePhysicalPathAuth: 'WebsiteUserPassThrough'
+        //                  AddBinding: true
+        //                  CreateOrUpdateAppPoolForWebsite: true
+        //                  ConfigureAuthenticationForWebsite: true
+        //                  AppPoolNameForWebsite: 'Spark'
+        //                  DotNetVersionForWebsite: 'v4.0'
+        //                  PipeLineModeForWebsite: 'Integrated'
+        //                  AppPoolIdentityForWebsite: 'ApplicationPoolIdentity'
+        //                  AnonymousAuthenticationForWebsite: true
+        //                  WindowsAuthenticationForWebsite: false
+        //                  protocol: 'http' 
+        //                  iPAddress: 'All Unassigned'
+        //                  port: '80'
+        //";
 
-//            //Act
-//            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
+        //            //Act
+        //            ConversionResponse gitHubOutput = conversion.ConvertAzurePipelineTaskToGitHubActionTask(yaml);
 
-//            //Assert
-//            string expected = @"
-//";
-//            expected = UtilityTests.TrimNewLines(expected);
-//            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
-//        }
+        //            //Assert
+        //            string expected = @"
+        //";
+        //            expected = UtilityTests.TrimNewLines(expected);
+        //            Assert.AreEqual(expected, gitHubOutput.actionsYaml);
+        //        }
 
         [TestMethod]
         public void SQLAzureDacPacDeployStepTest()
@@ -1415,7 +1441,7 @@ namespace AzurePipelinesToGitHubActionsConverter.Tests
             expected = UtilityTests.TrimNewLines(expected);
             Assert.AreEqual(expected, gitHubOutput.actionsYaml);
         }
-        
+
         [TestMethod]
         public void ArmTemplateDeploymentStepTest()
         {
